remove several warnings (Johannes Stezenbach)

+++ contrib/elftrunc.c
@@ -121,8 +121,8 @@ void trunc64(int fd)
   {
     if ((n=read(in,buf,sizeof(buf)))>0)
     {
-      write(out,buf,(size_t)((len<(size_t)n)?len:n));
-      len-=(len<(size_t)n)?len:n;
+      write(out,buf,(size_t)((len<(size_t)n)?len:(size_t)n));
+      len-=(len<(size_t)n)?len:(size_t)n;
     } else die(2,"read error");
   }
